file-submission-to-airflow/pyproject.toml
[project]
name = "airflow-file-submission"
version = "0.1.0"
description = "Proof of concept for submitting an external file to an Airflow orchestrator"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.3",
    "pandas>=2.3.3",
    "pandera[pandas]>=0.27.0",
    "pyarrow>=22.0.0",
    "requests>=2.31.0",
]
